bot/requirements.txt
requests==2.25.1
requests-cache==1.0.1
discord.py==2.2.3
mcuuid==1.1.0
Pillow==9.4.0
psutil==5.9.4
python-dateutil==2.8.2
numpy==1.24.1
python-dotenv==1.0.0